Ajustes em todos os CRUDS

+++ src/main/java/web/controlevacinacao/filter/ProdutoFilter.java
@@ -8,70 +8,89 @@ public class ProdutoFilter {
     private String nome;
     private Categoria categoria;
     private Tipo tipo;
-    private Integer minimoCusto;
-    private Integer maximoCusto;
-    private Integer minimoPrecoVenda;
-    private Integer maximoPrecoVenda;
+    private Double minimoCusto;
+    private Double maximoCusto;
+    private Double minimoPrecoVenda;
+    private Double maximoPrecoVenda;
     private Long estoque;
     private String fornecedor;
-    
+
     public Long getId() {
         return id;
     }
+
     public void setId(Long id) {
         this.id = id;
     }
+
     public String getNome() {
         return nome;
     }
+
     public void setNome(String nome) {
         this.nome = nome;
     }
+
     public Categoria getCategoria() {
         return categoria;
     }
+
     public void setCategoria(Categoria categoria) {
         this.categoria = categoria;
     }
+
     public Tipo getTipo() {
         return tipo;
     }
+
     public void setTipo(Tipo tipo) {
         this.tipo = tipo;
     }
-    public Integer getMinimoCusto() {
+
+    public Double getMinimoCusto() {
         return minimoCusto;
     }
-    public void setMinimoCusto(Integer minimoCusto) {
+
+    public void setMinimoCusto(Double minimoCusto) {
         this.minimoCusto = minimoCusto;
     }
-    public Integer getMaximoCusto() {
+
+    public Double getMaximoCusto() {
         return maximoCusto;
     }
-    public void setMaximoCusto(Integer maximoCusto) {
+
+    public void setMaximoCusto(Double maximoCusto) {
         this.maximoCusto = maximoCusto;
     }
-    public Integer getMinimoPrecoVenda() {
+
+    public Double getMinimoPrecoVenda() {
         return minimoPrecoVenda;
     }
-    public void setMinimoPrecoVenda(Integer minimoPrecoVenda) {
+
+    public void setMinimoPrecoVenda(Double minimoPrecoVenda) {
         this.minimoPrecoVenda = minimoPrecoVenda;
     }
-    public Integer getMaximoPrecoVenda() {
+
+    public Double getMaximoPrecoVenda() {
         return maximoPrecoVenda;
     }
-    public void setMaximoPrecoVenda(Integer maximoPrecoVenda) {
+
+    public void setMaximoPrecoVenda(Double maximoPrecoVenda) {
         this.maximoPrecoVenda = maximoPrecoVenda;
     }
+
     public Long getEstoque() {
         return estoque;
     }
+
     public void setEstoque(Long estoque) {
         this.estoque = estoque;
     }
+
     public String getFornecedor() {
         return fornecedor;
     }
+
     public void setFornecedor(String fornecedor) {
         this.fornecedor = fornecedor;
     }
